requirements.txt
Django==1.9.5
MySQL-python==1.2.5
argparse==1.2.1
django-crispy-forms==1.6.0
django-suit==0.2.18
djangorestframework==3.3.3
wsgiref==0.1.2
Django==1.9.5
MySQL-python==1.2.5
Pygments==2.1.3
argparse==1.2.1
django-crispy-forms==1.6.0
django-suit==0.2.18
djangorestframework==3.3.3
wsgiref==0.1.2
